add the decordation wires
lol the shaders were borked the whole time and I didn't realize it at all

+++ data/shaders/colorReplace.glsl
@@ -5,7 +5,7 @@ uniform float u_ncolors;
 vec4 frag(vec2 pos, vec2 uv, vec4 color, sampler2D tex) {
     vec4 o_color = texture2D(tex, uv);
     for(int i = 0; i < MAX_REPL; i++) {
-        if(i >= float(u_ncolors))
+        if(i >= int(u_ncolors))
             break;
         if(distance(o_color.rgb, u_colors_from[i].rgb / 255.) < .01) {
             if(distance(u_colors_from[i].rgb, vec3(0.)) < .01)
